+++ mesos/mesos.go
@@ -98,9 +98,8 @@ func (e *Mesos) Kill(taskID string, agentID string) error {
 func (e *Mesos) Call(message *mesosproto.Call) error {
 	message.FrameworkID = e.Framework.FrameworkInfo.ID
 	body, err := marshaller.MarshalToString(message)
-
 	if err != nil {
-		logrus.WithField("func", "mesos.Call").Debug("Could not Marshal message:", err.Error())
+		logrus.WithField("func", "mesos.Call").Error("Could not Marshal message:", err.Error())
 		return err
 	}